requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==12.1.1
pytz==2026.1.post1
sqlparse==0.5.5
